Bug ?
concordium-client config account import <path/to/exported/file> --name bakerAccount

Even if we give the wrong path to the wallet before using this code, when I use the code, it imports the wallet, but it gives the error does not chain. this error is eliminated when the file path is corrected.