B2 Submission Andrste
when the problem occurred with concordium, my node started to have issues. So my B1 failed with this server and B2 also didn't bake through all this timeframe when issue occurred... I needed to restart for a new version of concordium and flush the data. I hope you will consider this into review. Thank you.