B2 Logs
Ok so on 22th of February I did restarted and flushed my server, it showed data on concordium dashboard. But I see now it again failed... so that happened the same at 5th of February after that issue with concordium I did restarted on 12th of February but not flushed the DB and it was running fine, showed data on dashboard. Came on 22th of February it just showed: Unknown node ID: 2911d64eca809cac. You will see from the B2 logs which supposed to be also and B1 logs that on 22th when I restarted it recorded that day that it was running... now it is 24th same situation it is not running ... You will see that I started node on 2021-01-18... I hope you will consider this issue.